requirements.txt
# anyio==3.6.2
# appdirs==1.4.4
# argon2-cffi==21.3.0
# argon2-cffi-bindings==21.2.0
# arrow==1.2.3
# asttokens==2.2.1
# attrs==23.1.0
# backcall==0.2.0
# beautifulsoup4==4.12.2
# bleach==6.0.0
# brotlipy==0.7.0
# certifi @ file:///work/ci_py311/certifi_1676822443922/work/certifi
# cffi @ file:///work/ci_py311/cffi_1676822533496/work
# charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
# comm==0.1.3
# contourpy==1.0.7
# cryptography @ file:///work/ci_py311_2/cryptography_1679335572456/work
# cycler==0.11.0
# debugpy==1.6.7
# decorator==5.1.1
# defusedxml==0.7.1
# executing==1.2.0
# fastjsonschema==2.16.3
# fonttools==4.39.3
# fqdn==1.5.1
# htmlmin==0.1.12
# idna @ file:///work/ci_py311/idna_1676822698822/work
# ImageHash==4.3.1
# ipykernel==6.22.0
# ipython==8.12.0
# ipython-genutils==0.2.0
# ipywidgets==8.0.6
# isoduration==20.11.0
# jedi==0.18.2
# Jinja2==3.1.2
# joblib @ file:///work/ci_py311/joblib_1676837020463/work
# jsonpointer==2.3
# jsonschema==4.17.3
# jupyter-contrib-core==0.4.2
# jupyter-contrib-nbextensions==0.7.0
# jupyter-events==0.6.3
# jupyter-highlight-selected-word==0.2.0
# jupyter-nbextensions-configurator==0.6.1
# jupyter_client==8.2.0
# jupyter_core==5.3.0
# jupyter_server==2.5.0
# jupyter_server_terminals==0.4.4
# jupyterlab-pygments==0.2.2
# jupyterlab-widgets==3.0.7
# kiwisolver==1.4.4
# lxml==4.9.2
# MarkupSafe==2.1.2
# matplotlib==3.6.3
# matplotlib-inline==0.1.6
# mistune==2.0.5
# mkl-fft==1.3.1
# mkl-random @ file:///work/ci_py311/mkl_random_1676828032909/work
# mkl-service==2.4.0
# multimethod==1.9.1
# nbclassic==0.5.5
# nbclient==0.7.3
# nbconvert==7.3.1
# nbformat==5.8.0
# nest-asyncio==1.5.6
# networkx==3.1
# notebook==6.5.4
# notebook_shim==0.2.2
# numpy @ file:///work/ci_py311/numpy_and_numpy_base_1676823870601/work
# packaging @ file:///croot/packaging_1678965309396/work
# pandas==1.5.3
# pandocfilters==1.5.0
# parso==0.8.3
# patsy==0.5.3
# pexpect==4.8.0
# phik==0.12.3
# pickleshare==0.7.5
# Pillow==9.5.0
# platformdirs==3.2.0
# pooch @ file:///tmp/build/80754af9/pooch_1623324770023/work
# prometheus-client==0.16.0
# prompt-toolkit==3.0.38
# psutil==5.9.5
# ptyprocess==0.7.0
# pure-eval==0.2.2
# pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
# pydantic==1.10.7
# Pygments==2.15.1
# pyOpenSSL @ file:///croot/pyopenssl_1678965284384/work
# pyparsing==3.0.9
# pyrsistent==0.19.3
# PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
# python-dateutil==2.8.2
# python-json-logger==2.0.7
# pytz==2023.3
# PyWavelets==1.4.1
# PyYAML==6.0
# pyzmq==25.0.2
# requests @ file:///work/ci_py311_2/requests_1679337615230/work
# rfc3339-validator==0.1.4
# rfc3986-validator==0.1.1
# scikit-learn @ file:///croot/scikit-learn_1680198742267/work
# scipy==1.9.3
# seaborn==0.12.2
# Send2Trash==1.8.0
# six @ file:///tmp/build/80754af9/six_1644875935023/work
# sniffio==1.3.0
# soupsieve==2.4.1
# stack-data==0.6.2
# statsmodels==0.13.5
# tangled-up-in-unicode==0.2.0
# terminado==0.17.1
# threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
# tinycss2==1.2.1
# tornado==6.3.1
# tqdm==4.64.1
# traitlets==5.9.0
# typeguard==2.13.3
# typing_extensions==4.5.0
# tzdata==2023.3
# uri-template==1.2.0
# urllib3 @ file:///croot/urllib3_1680254681959/work
# visions==0.7.5
# wcwidth==0.2.6
# webcolors==1.13
# webencodings==0.5.1
# websocket-client==1.5.1
# widgetsnbextension==4.0.7
# ydata-profiling==4.1.2
# name: my_environment
# channels:
#   - conda-forge

scikit-learn
pandas
matplotlib
azureml-sdk


